+++ djangocms_slick_slider/models.py
@@ -1,10 +1,11 @@
 # -*- coding: utf-8 -*-
 from __future__ import unicode_literals
-
 from django.db import models
+from django.dispatch import receiver
+from django.db.models.signals import pre_save
 from django.utils.encoding import python_2_unicode_compatible
 from django.utils.translation import ugettext_lazy as _
-
+from djangocms_text_ckeditor.fields import HTMLField
 from cms.models.pluginmodel import CMSPlugin
 from filer.fields.image import FilerImageField
 from jsonfield import JSONField
@@ -35,12 +36,27 @@ class SlickSlider(CMSPlugin):
             'Use JSON format and check the errors in the editor<br>'
             'You can also use online JSON validators'))
 
+    speed = models.PositiveIntegerField(
+        verbose_name=_('Speed (ms)'),
+        default=300)
+
+    slides_to_show = models.PositiveIntegerField(
+        verbose_name=_('Slides to show'),
+        default=1)
+
+    arrows = models.BooleanField(
+        verbose_name=_('Arrows'),
+        default=True)
+
     arrow_color = models.CharField(
         verbose_name=_('arrow color'),
         max_length=255,
         default="#666",
         help_text=_('Define the color of slider arrows here. All CSS '
                     'color values work (e.g. #efefef).'))
+    dots = models.BooleanField(
+        _('Dots'),
+        default=True)
 
     def copy_relations(self, oldinstance):
         """
@@ -57,6 +73,17 @@ class SlickSlider(CMSPlugin):
         String representation of SlickSlider class.
         """
         return "{title}".format(title=self.title)
+
+
+@receiver(pre_save, sender=SlickSlider)
+def settings_handler(sender, instance, **kwargs):
+    """ Update `instance.settings` with slide settings form fields
+    """
+    instance.settings['speed'] = instance.speed
+    instance.settings['arrows'] = instance.arrows
+    instance.settings['dots'] = instance.dots
+    instance.settings['slidesToShow'] = instance.slides_to_show
+    return
 
 
 @python_2_unicode_compatible
@@ -85,7 +112,7 @@ class SlickSliderImage(models.Model):
         help_text=_('open link in new window'),
         default=True)
 
-    caption_text = models.TextField(
+    caption_text = HTMLField(
         _('caption text'),
         null=True,
         blank=True)